+++ myokit/formats/cellml/parser_1.py
@@ -44,7 +44,7 @@ class CellMLParsingError(myokit.ImportError):
     """
     def __init__(self, message, element=None):
         if element is not None:
-            try:
+            try:    # pragma: no cover
                 line = str(element.sourceline)
                 message = 'Error on line ' + line + '. ' + message
             except AttributeError:
